Move pandas_not_installed fixture to conftest.py
Since tests are distributed in multiple folders, putting conftest.py
into the package directory.
