+++ skops/card/tests/test_card.py
@@ -3,7 +3,6 @@ import os
 import pickle
 import tempfile
 from pathlib import Path
-from unittest.mock import patch
 
 import matplotlib.pyplot as plt
 import numpy as np
@@ -391,14 +390,8 @@ class TestTableSection:
         with pytest.raises(ValueError, match=msg):
             TableSection(table=table)
 
-    def test_pandas_not_installed(self, table_dict):
-        # patch import so that it raises an ImportError when trying to import
-        # pandas. This works because pandas is only imported lazily.
-        def mock_import(name, *args, **kwargs):
-            if name == "pandas":
-                raise ImportError
-            return __import__(name, *args, **kwargs)
-
-        with patch("builtins.__import__", side_effect=mock_import):
-            section = TableSection(table=table_dict)
-            assert section._is_pandas_df is False
+    def test_pandas_not_installed(self, table_dict, pandas_not_installed):
+        # use pandas_not_installed fixture from conftest.py to pretend that
+        # pandas is not installed
+        section = TableSection(table=table_dict)
+        assert section._is_pandas_df is False
